+++ src/main/java/org/jointheleague/jcodrone/TestPorts.java
@@ -12,10 +12,10 @@ public class TestPorts {
             Thread.sleep(200);
 //            drone.sendCommand(CommandType.FLIGHT_EVENT, FlightEvent.TAKEOFF.value());
             Thread.sleep(2000);
-            LightMode mode = new LightModeBuilder().setColor("BLUE").setInterval((short) 5).setMode(LightModeDrone.ARM_HOLD).build();
+            LightMode mode = new LightModeBuilder().setColor("BLUE").setInterval(5).setMode(LightModeDrone.ARM_HOLD).build();
             drone.lightMode(mode);
             Thread.sleep(1000);
-            mode = new LightModeBuilder().setColor("green").setInterval((short) 200).setMode(LightModeDrone.EYE_DIMMING).build();
+            mode = new LightModeBuilder().setColor("green").setInterval(200).setMode(LightModeDrone.EYE_DIMMING).build();
             drone.lightMode(mode);
             Thread.sleep(3000);
 //            drone.sendCommand(CommandType.FLIGHT_EVENT, FlightEvent.LANDING.value());
